Alteração do tipo do atributo votes de long para Double Criação da classe MovieDTO. E mudança no atributo score. Tentativa de gravação do arquivo temporário;

+++ src/main/java/br/com/letscode/java/moviebattleapi/movie/MovieRestRepository.java
@@ -61,7 +61,7 @@ public class MovieRestRepository {
         movie.setTitle(token.nextToken());
         movie.setYear(Integer.valueOf(token.nextToken()));
         movie.setRating(Double.valueOf(token.nextToken()));
-        movie.setVotes(Long.valueOf(token.nextToken()));
+        movie.setVotes(Double.valueOf(token.nextToken()));
         return movie;
 
     }
